+++ test/injector_test.dart
@@ -1,6 +1,4 @@
-import 'dart:collection';
-
-import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:provide_it/src/injector/injector.dart';
 import 'package:provide_it/src/injector/param.dart';
@@ -243,38 +241,40 @@ void main() {
     test('should manually provide some parameters', () {
       final injector = Injector(({String? a = '', int? b}) => (a, b));
       expect(injector(), ('', null));
-      expect(injector({#a: 'a'}), ('a', null));
-      expect(injector({#b: 1}), ('', 1));
-      expect(injector({#a: 'a', #b: 1}), ('a', 1));
+      expect(injector({'a': 'a'}), ('a', null));
+      expect(injector({'b': 1}), ('', 1));
+      expect(injector({'a': 'a', 'b': 1}), ('a', 1));
     });
 
-    test('should resolve parameters resolution', () {
-      Object? call(Map<Object, dynamic> parameters, Param param) {
-        Object? arg;
+    test('should inject by name, position or type', () {
+      final injector = Injector(Text.new, parameters: {
+        '0': 'Hello',
+        'style': TextStyle(),
+        '$TextAlign': TextAlign.center,
+      });
 
-        parameters = HashMap(
-          equals: (a, b) =>
-              a is Type && b is String ? a.toString() == b : a == b,
-          hashCode: (key) =>
-              key is Type ? key.toString().hashCode : key.hashCode,
-        )..addAll(parameters);
+      final result = injector();
+      expect(result, isA<Text>());
 
-        arg ??= parameters[param.symbol] ??
-            parameters[param.index] ??
-            parameters[param.type] ??
-            parameters['${param.name ?? param.index}'];
+      final text = result as Text;
+      expect(text.data, 'Hello');
+      expect(text.style, isA<TextStyle>());
+      expect(text.textAlign, TextAlign.center);
+    });
 
-        return arg;
-      }
+    test('should inject positional & default typed parameters', () {
+      final injector = Injector(DateTime.new, parameters: {
+        '0': 2018, // by position
+        '1': 12,
+        '2': 4,
+        '3': 18,
+        '4': 30,
+        '$int': 60, // by type
+      });
 
-      final param = PositionalParam('String',
-          index: 3, isRequired: true, owner: Injector(() {}));
-
-      final result = call({3: 'by index'}, param);
-      expect(result, 'by index');
-
-      final result2 = call({String: 'by type'}, param);
-      expect(result2, 'by type');
+      final result = injector();
+      expect(result, isA<DateTime>());
+      expect(result, DateTime(2018, 12, 4, 18, 30, 60, 60, 60));
     });
   });
 }